+++ dashboard/views/__init__.py
@@ -1,24 +1,7 @@
-from dashboard.views.__pain_and_suffering import *
-
-'''
-import csv
-import random
-
-from django.contrib import messages, auth
-from django.urls import reverse, reverse_lazy
-from django.db.models import Q
-from django.http import HttpResponseRedirect, HttpResponse
-from django.shortcuts import render, get_object_or_404, redirect
-from django.views.generic import *
-from django.views.generic.edit import UpdateView, DeleteView
-from django.db import transaction
-from django.conf import settings
-from django.contrib.auth.decorators import login_required
-from django.core.mail import send_mail
-from django.core.exceptions import ObjectDoesNotExist
-
-
-from dashboard.utils import *
-from datetime import datetime
-from dashboard.forms import *
-'''
+from dashboard.views._auth import *
+from dashboard.views._brother import *
+from dashboard.views._committee import *
+from dashboard.views._general import *
+from dashboard.views._home import *
+from dashboard.views._house_management import *
+from dashboard.views._positions import *
