+++ dashboard/views/_auth.py
@@ -1,6 +1,7 @@
 from django.contrib import auth, messages
-from django.urls import reverse
 from django.http import HttpResponseRedirect
+from django.urls import reverse
+from django.views.generic import View
 
 from dashboard.forms import ChangePasswordForm
