+++ dashboard/views/_positions/_non_ec/_membership_development_chair.py
@@ -1,0 +1,9 @@
+from django.shortcuts import render
+
+from dashboard.utils import committee_meeting_panel,verify_position
+
+@verify_position(['Membership Development Chair', 'Vice President', 'President', 'Adviser'])
+def memdev_c(request):
+    committee_meetings, context = committee_meeting_panel('Membership Development Chair')
+
+    return render(request, 'memdev-chair.html', context)
